+++ frontend/src/routes/+layout.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
-  import { page } from '$app/stores';
+  import { browser } from '$app/environment';
+  import { goto } from '$app/navigation';
   import { verifyId } from '$lib/api/admin.api';
   import GlobalLogin from '$lib/components/shared/GlobalLogin.svelte';
   import Header from '$lib/components/shared/Header.svelte';
@@ -21,6 +22,7 @@
   import { onMount } from 'svelte';
 
   $: isVerified = $identifierStore ? true : false;
+  $: if (!isVerified && browser) goto('/');
 
   onMount(async () => {
     if ($identifierStore) {
@@ -35,7 +37,6 @@
     }
   });
   $: locale = $languageStore;
-  $: isAdminRoute = $page.url.pathname.replace('/', '') === 'admin';
   $: theme = $themeStore;
   $: document.documentElement.setAttribute('theme', theme);
   $: getAppClasses = `app background_${$backgroundStore}_${$themeStore === darkThemeValue ? 'dark' : 'light'}`;
@@ -60,11 +61,13 @@
     </div>
   {/if}
 
-  <Header {locale} />
+  {#if isVerified}
+    <Header {locale} />
+  {/if}
 
   {#if !isVerified}
     <div class="login_container">
-      <GlobalLogin {isVerified} isAdminLogin={isAdminRoute} />
+      <GlobalLogin {isVerified} />
     </div>
   {:else}
     <main>
@@ -118,8 +121,8 @@
     display: flex;
     align-items: center;
     justify-content: center;
-    height: 80vh;
-    height: 80dvh;
+    height: 100vh;
+    height: 100dvh;
   }
 
   .confetti_container {